Re-structure, improve entity row structs
