+++ src/panels/lovelace/editor/types.ts
@@ -1,15 +1,4 @@
 import {
-  array,
-  boolean,
-  dynamic,
-  enums,
-  literal,
-  object,
-  optional,
-  string,
-  union,
-} from "superstruct";
-import {
   ActionConfig,
   LovelaceCardConfig,
   LovelaceViewConfig,
@@ -92,82 +81,3 @@ export interface SubElementEditorConfig {
 export interface EditSubElementEvent {
   subElementConfig: SubElementEditorConfig;
 }
-
-export const actionConfigStruct = dynamic((value: any) => {
-  if (value && value.action in actionConfigMap) {
-    return actionConfigMap[value.action];
-  }
-  return actionConfigStructType;
-});
-
-const actionConfigStructUser = object({
-  user: string(),
-});
-
-const actionConfigStructConfirmation = union([
-  boolean(),
-  object({
-    text: optional(string()),
-    excemptions: optional(array(actionConfigStructUser)),
-  }),
-]);
-
-const actionConfigStructUrl = object({
-  action: literal("url"),
-  url_path: string(),
-  confirmation: optional(actionConfigStructConfirmation),
-});
-
-const actionConfigStructService = object({
-  action: literal("call-service"),
-  service: string(),
-  service_data: optional(object()),
-  target: optional(
-    object({
-      entity_id: optional(union([string(), array(string())])),
-      device_id: optional(union([string(), array(string())])),
-      area_id: optional(union([string(), array(string())])),
-    })
-  ),
-  confirmation: optional(actionConfigStructConfirmation),
-});
-
-const actionConfigStructNavigate = object({
-  action: literal("navigate"),
-  navigation_path: string(),
-  confirmation: optional(actionConfigStructConfirmation),
-});
-
-const actionConfigMap = {
-  url: actionConfigStructUrl,
-  navigate: actionConfigStructNavigate,
-  "call-service": actionConfigStructService,
-};
-
-export const actionConfigStructType = object({
-  action: enums([
-    "none",
-    "toggle",
-    "more-info",
-    "call-service",
-    "url",
-    "navigate",
-  ]),
-  confirmation: optional(actionConfigStructConfirmation),
-});
-
-export const entitiesConfigStruct = union([
-  object({
-    entity: string(),
-    name: optional(string()),
-    icon: optional(string()),
-    image: optional(string()),
-    secondary_info: optional(string()),
-    format: optional(string()),
-    state_color: optional(boolean()),
-    tap_action: optional(actionConfigStruct),
-    hold_action: optional(actionConfigStruct),
-    double_tap_action: optional(actionConfigStruct),
-  }),
-  string(),
-]);
